refactor: service 참조 수정
