+++ src/main/java/roomescape/reservation/service/services/ReservationCommonService.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.service;
+package roomescape.reservation.service.services;
 
 import java.time.LocalDate;
 import org.springframework.stereotype.Service;
